implement valid_isbn13? - first try which was bad
bad because i was writing out algorithms before writing out the tests.
i was able to accomplish this only because the algo was very simple and
the whole thing could be kept in my head. but i doubt this same process
would work for a more complicated project.

will try again with real TDD